Gestion de IPR
Former-commit-id: f40da04ca324fa46780803204add5f7ea7664c76 [formerly 7043e5eb05ac269af2188114ff15276f5d124908]
Former-commit-id: 2eb203675ba72db8e51fef7a9d0623cc26791127

+++ server/sql/bhima.sql
@@ -1384,7 +1384,7 @@ create table `hollyday` (
 drop table if exists `taxe_ipr`;
 create table `taxe_ipr` (
   id                      int unsigned auto_increment not null,
-  taux                    int not null,
+  taux                    float not null,
   tranche_annuelle_debut  float,
   tranche_annuelle_fin    float,
   tranche_mensuelle_debut float,
